all basic functionality on the SMSGateway
Made a little test java server. Phone will send incoming texts to the
server in format [<phonenumber>]<message>|. server sent it back and I
replied to a friend with the same message. Need to test it with a
non-java server to make sure my networking is right, though I did
convert my strings to UTF-8 so I think it's good.