ACTUAL TEST CASES
We have actual test cases for the device hub related to GPS requests.
They seem to be working, you have to manually check the texts sent on a
phone, but the test cases do make sure there is a response on the
channel included in the request. I haven't tested processing requests
with malformed payloads, so for now I'm going to rely on Leo continuing
to produce bug free code.